add create user transaction (#56)
