+++ db/sqlc/store.go
@@ -12,6 +12,8 @@ import (
 type Store interface {
 	Querier
 	TransferTx(ctx context.Context, arg TransferTxParams) (TransferTxResult, error)
+	CreateUserTx(ctx context.Context, arg CreateUserTxParams) (
+		CreateUserTxResult, error)
 }
 
 type SQLStore struct {
@@ -46,99 +48,4 @@ func (store *SQLStore) execTx(ctx context.Context, fn func(*Queries) error) erro
 	}
 
 	return tx.Commit()
-}
-
-type TransferTxParams struct {
-	FromAccountID int64 `json:"from_account_id"`
-	ToAccountID   int64 `json:"to_account_id"`
-	Amount        int64 `json:"amount"`
-}
-
-type TransferTxResult struct {
-	Transfer    Transfer `json:"transfer"`
-	FromAccount Account  `json:"from_account"`
-	ToAccount   Account  `json:"to_account"`
-	FromEntry   Entry    `json:"from_entry"`
-	ToEntry     Entry    `json:"to_entry"`
-}
-
-// TransferTx は口座から口座への送金を行う
-func (store *SQLStore) TransferTx(ctx context.Context, arg TransferTxParams) (
-	TransferTxResult, error) {
-	var result TransferTxResult
-
-	err := store.execTx(ctx, func(q *Queries) error {
-		var err error
-
-		// 1. transfer tableにInsert
-
-		result.Transfer, err = q.CreateTransfer(ctx, CreateTransferParams(arg))
-		if err != nil {
-			return err
-		}
-
-		// 2. 口座Aに 送信した料金分マイナスの entryをInsert
-
-		result.FromEntry, err = q.CreateEntry(ctx, CreateEntryParams{
-			AccountID: arg.FromAccountID,
-			Amount:    -arg.Amount,
-		})
-
-		if err != nil {
-			return err
-		}
-
-		// 3. 口座Bに受け取った料金分プラスするentryをInsert
-
-		result.ToEntry, err = q.CreateEntry(ctx, CreateEntryParams{
-			AccountID: arg.ToAccountID,
-			Amount:    arg.Amount,
-		})
-
-		if err != nil {
-			return err
-		}
-
-		if arg.FromAccountID < arg.ToAccountID {
-			result.FromAccount, result.ToAccount, err = addMoney(ctx, q,
-				arg.FromAccountID, -arg.Amount,
-				arg.ToAccountID, arg.Amount)
-
-		} else {
-			result.ToAccount, result.FromAccount, err = addMoney(ctx, q,
-				arg.ToAccountID, arg.Amount,
-				arg.FromAccountID, -arg.Amount)
-
-		}
-
-		return err
-	})
-
-	return result, err
-}
-
-func addMoney(
-	ctx context.Context,
-	q *Queries,
-	accountID1 int64,
-	amount1 int64,
-	accountID2 int64,
-	amount2 int64,
-) (account1, account2 Account, err error) {
-
-	account1, err = q.AddAccountBalance(ctx, AddAccountBalanceParams{
-		ID:     accountID1,
-		Amount: amount1,
-	})
-
-	if err != nil {
-		return
-	}
-
-	account2, err = q.AddAccountBalance(ctx, AddAccountBalanceParams{
-		ID:     accountID2,
-		Amount: amount2,
-	})
-
-	return
 }
